requirements.txt
fastapi[standard]
psycopg[binary]
itsdangerous
starlette
authlib
PyJWT

